backend/requirements.firestore-runtime.txt
# SAO Backend dependencies for Firestore-only runtime image
# Derived from requirements.txt, excluding SQL-only runtime tooling.

# Web Framework (runtime)
fastapi==0.135.1
python-multipart==0.0.22
uvicorn[standard]==0.30.0

# Validation & Serialization (runtime)
email-validator>=2.0.0
pydantic==2.7.4
pydantic-settings==2.1.0

# Authentication & Security (runtime)
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
PyJWT==2.9.0
python-jose[cryptography]==3.3.0

# Utilities / Integrations (runtime)
google-cloud-firestore==2.20.0
google-cloud-storage==2.18.2
numpy==1.26.4
pillow==10.4.0
pypdf==5.3.1
pypdfium2==4.30.0
python-dotenv==1.0.0
rapidocr-onnxruntime==1.3.24

# ASGI toolkit (runtime, explicit pin for FastAPI compatibility)
starlette==0.52.1
